database connection overhaul until VoiceStateUpdateService

+++ src/Services/ReminderService.py
@@ -5,6 +5,7 @@ import discord
 from discord import Member
 
 from src.Helper.CreateNewDatabaseConnection import getDatabaseConnection
+from src.Helper.Database import Database
 from src.Helper.DictionaryFuntionKeyDecorator import validateKeys
 from src.Helper.SendDM import sendDM
 from src.Helper.WriteSaveQuery import writeSaveQuery
@@ -18,7 +19,11 @@ logger = logging.getLogger("KVGG_BOT")
 class ReminderService:
 
     def __init__(self, client: discord.Client):
-        self.databaseConnection = getDatabaseConnection()
+        """
+        :param client:
+        :raise ConnectionError:
+        """
+        self.database = Database()
         self.client = client
 
     @validateKeys
@@ -92,49 +97,45 @@ class ReminderService:
 
                     return "Es ist ein Fehler aufgetreten!"
 
-        if not (dcUserDb := getDiscordUser(self.databaseConnection, member)):
+        if not (dcUserDb := getDiscordUser(member)):
             logger.debug("cant proceed, no DiscordUser")
 
             return "Es gab ein Problem!"
 
         if whatsapp:
-            with self.databaseConnection.cursor() as cursor:
-                query = "SELECT * " \
-                        "FROM whatsapp_setting " \
-                        "WHERE discord_user_id = %s"
+            query = "SELECT * " \
+                    "FROM whatsapp_setting " \
+                    "WHERE discord_user_id = %s"
+            data = self.database.queryOneResult(query, (dcUserDb['id'],))
 
-                cursor.execute(query, (dcUserDb['id'],))
+            if not data:
+                logger.debug("User cannot receive whatsapp notifications")
 
-                if not (data := cursor.fetchone()):
-                    logger.debug("User cannot receive whatsapp notifications")
+                answerAppendix = "Allerdings kannst du keine Whatsapp-Benachrichtigungen bekommen."
+                whatsapp = False
+            else:
+                logger.debug("user registered for whatsapp reminder as well")
 
-                    answerAppendix = "Allerdings kannst du keine Whatsapp-Benachrichtigungen bekommen."
-                    whatsapp = False
-                else:
-                    logger.debug("user registered for whatsapp reminder as well")
+                whatsapp = True
 
-                    whatsapp = True
-
-                # delete unnecessary data overhead
-                del data
+            # delete unnecessary data overhead
+            del data
         else:
             whatsapp = False
 
-        with self.databaseConnection.cursor() as cursor:
-            query = "INSERT INTO reminder " \
-                    "(discord_user_id, content, time_to_sent, sent_at, whatsapp, repeat_in_minutes) " \
-                    "VALUES (%s, %s, %s, %s, %s, %s)"
+        query = "INSERT INTO reminder " \
+                "(discord_user_id, content, time_to_sent, sent_at, whatsapp, repeat_in_minutes) " \
+                "VALUES (%s, %s, %s, %s, %s, %s)"
 
-            cursor.execute(query,
-                           (dcUserDb['id'],
-                            content,
-                            date,
-                            None,
-                            whatsapp,
-                            minutesLeft if repeatTime and repeatType else None), )
-            self.databaseConnection.commit()
-
-        logger.debug("saved new reminder to database")
+        if self.database.runQueryWithoutFetching(query, (dcUserDb['id'],
+                                                         content,
+                                                         date,
+                                                         None,
+                                                         whatsapp,
+                                                         minutesLeft if repeatTime and repeatType else None,)):
+            logger.debug("saved new reminder to database")
+        else:
+            return "Es gab ein Problem beim speicher des Reminders."
 
         return "Deine Erinnerung wurde erfolgreich gespeichert! " + \
             (answerAppendix if 'answerAppendix' in locals() else "")
@@ -147,20 +148,16 @@ class ReminderService:
         :param member: Member, who asked for his / her remainders
         :return:
         """
-        with self.databaseConnection.cursor() as cursor:
-            query = "SELECT * " \
-                    "FROM reminder " \
-                    "WHERE (SELECT id FROM discord WHERE user_id = %s) = discord_user_id " \
-                    "and time_to_sent IS NOT NULL"
+        query = "SELECT * " \
+                "FROM reminder " \
+                "WHERE (SELECT id FROM discord WHERE user_id = %s) = discord_user_id " \
+                "and time_to_sent IS NOT NULL"
+        reminders = self.database.queryAllResults(query, (int(member.id),))
 
-            cursor.execute(query, (int(member.id),))
+        if not reminders:
+            logger.debug("reminders for %s were empty" % member.name)
 
-            if not (data := cursor.fetchall()):
-                logger.debug("reminders for %s were empty" % member.name)
-
-                return "Du hast keine aktiven Reminders."
-
-            reminders = [dict(zip(cursor.column_names, date)) for date in data]
+            return "Du hast keine aktiven Reminders."
 
         answer = "Du hast folgende Reminder: (die vorderen Zahlen sind die individuellen IDs)\n\n"
 
@@ -181,19 +178,15 @@ class ReminderService:
 
         :return:
         """
-        with self.databaseConnection.cursor() as cursor:
-            query = "SELECT r.*, d.user_id " \
-                    "FROM reminder r INNER JOIN discord d on r.discord_user_id = d.id " \
-                    "WHERE time_to_sent is NOT NULL"
+        query = "SELECT r.*, d.user_id " \
+                "FROM reminder r INNER JOIN discord d on r.discord_user_id = d.id " \
+                "WHERE time_to_sent is NOT NULL"
+        reminders = self.database.queryAllResults(query)
 
-            cursor.execute(query)
+        if not reminders:
+            logger.debug("no reminders were found")
 
-            if not (data := cursor.fetchall()):
-                logger.debug("no reminders were found")
-
-                return
-
-            reminders = [dict(zip(cursor.column_names, date)) for date in data]
+            return
 
         tempReminders = []
 
@@ -205,16 +198,14 @@ class ReminderService:
 
         reminders = tempReminders
 
-        with self.databaseConnection.cursor() as cursor:
-            for reminder in reminders:
-                # remove column that does not exist in the reminder field list
-                reminder.pop('user_id', None)
+        for reminder in reminders:
+            # remove column that does not exist in the reminder field list
+            reminder.pop('user_id', None)
 
-                query, nones = writeSaveQuery("reminder", reminder['id'], reminder)
+            query, nones = writeSaveQuery("reminder", reminder['id'], reminder)
 
-                cursor.execute(query, nones)
-
-            self.databaseConnection.commit()
+            if not self.database.saveChangesToDatabase(query, nones):
+                logger.critical("couldn't save reminder into database, id: %s" % str(reminder['id']))
 
     @validateKeys
     def deleteReminder(self, member: Member, id: int) -> str:
@@ -233,30 +224,25 @@ class ReminderService:
 
             return "Bitte gib eine korrekte ID ein!"
 
-        with self.databaseConnection.cursor() as cursor:
-            query = "SELECT r.* " \
-                    "FROM reminder r INNER JOIN discord d ON r.discord_user_id = d.id " \
-                    "WHERE d.user_id = %s AND r.time_to_sent IS NOT NULL"
+        query = "SELECT r.* " \
+                "FROM reminder r INNER JOIN discord d ON r.discord_user_id = d.id " \
+                "WHERE d.user_id = %s AND r.time_to_sent IS NOT NULL"
 
-            cursor.execute(query, (member.id,))
+        reminders = self.database.queryAllResults(query, (member.id,))
 
-            if not (data := cursor.fetchall()):
-                logger.debug("no entries to delete")
+        if not reminders:
+            logger.debug("no entries to delete")
 
-                return "Du hast keine aktiven Reminders!"
-
-            reminders = [dict(zip(cursor.column_names, date)) for date in data]
+            return "Du hast keine aktiven Reminders!"
 
         if not any(reminder['id'] == id for reminder in reminders):
             logger.debug("id was not found in personal reminders")
 
             return "Du hast keinen Reminder mit dieser ID!"
 
-        with self.databaseConnection.cursor() as cursor:
-            query = "DELETE FROM reminder WHERE id = %s"
+        query = "DELETE FROM reminder WHERE id = %s"
 
-            cursor.execute(query, (id,))
-            self.databaseConnection.commit()
+        self.database.runQueryWithoutFetching(query, (id,))
 
         logger.debug("deleted entry from database")
         return "Dein Reminder wurde erfolgreich gelöscht."
@@ -276,20 +262,21 @@ class ReminderService:
             return reminder
 
         if reminder['whatsapp']:
-            with self.databaseConnection.cursor() as cursor:
-                query = "INSERT INTO message_queue (message, user_id, created_at, trigger_user_id, is_join_message) " \
-                        "VALUES (%s, " \
-                        "(SELECT id FROM user WHERE discord_user_id = " \
-                        "(SELECT id FROM discord WHERE user_id = %s LIMIT 1) LIMIT 1), " \
-                        "%s, " \
-                        "(SELECT id FROM discord WHERE user_id = %s LIMIT 1), " \
-                        "FALSE)"
+            query = "INSERT INTO message_queue (message, user_id, created_at, trigger_user_id, is_join_message) " \
+                    "VALUES (%s, " \
+                    "(SELECT id FROM user WHERE discord_user_id = " \
+                    "(SELECT id FROM discord WHERE user_id = %s LIMIT 1) LIMIT 1), " \
+                    "%s, " \
+                    "(SELECT id FROM discord WHERE user_id = %s LIMIT 1), " \
+                    "FALSE)"
 
-                cursor.execute(query, ("Hier ist deine Erinnerung:\n\n" + reminder['content'],
-                                       member.id,
-                                       datetime.now(),
-                                       member.id,))
-                self.databaseConnection.commit()
+            if not self.database.runQueryWithoutFetching(query,
+                                                         ("Hier ist deine Erinnerung:\n\n" + reminder['content'],
+                                                          member.id,
+                                                          datetime.now(),
+                                                          member.id,)):
+                logger.critical("couldn't save message into database")
+            else:
                 logger.debug("saved whatsapp into message queue")
 
         try:
